move MetricObject to metric.py
